fix(tuttid): keep ActivityProjection under revive file-length limit
Move runtime-node helpers and sessionIsChild out of activity_projection.go so changed Go lint stays under the 800-line revive cap.
