+++ services/tuttid/service/agent/activity_projection_reconcile.go
@@ -4,8 +4,10 @@ import (
 	"context"
 	"errors"
 	"log/slog"
+	"strings"
 
 	agenthost "github.com/tutti-os/tutti/packages/agent/host"
+	agentactivitybiz "github.com/tutti-os/tutti/packages/agent/store-sqlite"
 )
 
 // SettleStaleTurnsOnStartup is the daemon-start reconciliation of protocol v2
@@ -40,4 +42,19 @@ func (p *ActivityProjection) SettleStaleTurnsOnStartup(ctx context.Context) erro
 	}
 	p.observeCommittedOutsideHost(ctx, delta)
 	return nil
+}
+
+// sessionIsChild reports whether a canonical session is a provider-native
+// subagent session.
+func (p *ActivityProjection) sessionIsChild(ctx context.Context, workspaceID, agentSessionID string) bool {
+	workspaceID, agentSessionID = strings.TrimSpace(workspaceID), strings.TrimSpace(agentSessionID)
+	if p == nil || p.repo == nil || workspaceID == "" || agentSessionID == "" {
+		return false
+	}
+	session, found, err := p.repo.GetSession(ctx, workspaceID, agentSessionID)
+	if err != nil || !found {
+		return false
+	}
+	return strings.EqualFold(strings.TrimSpace(session.Kind), agentactivitybiz.SessionKindChild) ||
+		strings.TrimSpace(session.ParentToolCallID) != ""
 }
